Refactor films code, take controller out to the page component
