+++ src/const.js
@@ -1,6 +1,9 @@
 const MIN_PER_HOUR = 60;
 const RADIX_TEN = 10;
-const FILM_CARDS_PER_ROW = 5;
+const CardsPerRow = {
+  MAIN: 5,
+  EXTRA: 2,
+};
 const EMOJI = [`smile`, `sleeping`, `puke`, `angry`];
 
-export {MIN_PER_HOUR, RADIX_TEN, FILM_CARDS_PER_ROW, EMOJI};
+export {MIN_PER_HOUR, RADIX_TEN, CardsPerRow, EMOJI};
